Changed color_t support
Now functions always take just color_t and it can be taken to mean either a single data point or an array. It depends on the context. Make sure to carefully implement these functions and then hide them from the end user.

Also added a necessary pure virtual function that shifts a colot_t type to the appropriate address for a given number of pixels. (Depends on your implentation, assumes that the base color_t points to an array of your color type with a sufficient number of pixels worht of data) Whew that was a mouthful!