Added buffer mode
Buffer mode allows you to associate memory for each window, then draw to the memory without showing it on the screen. Then at a later time you can dump all the window's memory to the screen.

Note: in this process I realized that the implementation of hwFillFromArray had reversed the meaning of Vh. That is to say that Vh being true (vertical mode, where vertical lines are drawn first, i.e. pixels are drawn top to bottom, and columns are repeated left to right) was actually drawing rows of pixels right to left and repeating them down.

I discovered this b/c the buffer filling worked when using the derived class' fillFromArray function but not when only implementing hwpixel. I reversed the implementation in HyperDisplay to be more intuitive, but this could cause problems in any classes that expect the old meaning.